Version 3.0.0
This version is a significant improvement. It completely restructures
the codebase and adds a bunch of new features like modular BP's,
caches, devices, etc... TAGE BP is still not fully working but this is a
strong project for now. I'm sure there are bugs but it works with
optimizations now, floats, and atomics meaning more unique programs can
be run. This is the largest improvement yet for this project but there
still is a lot to be done for this.